+++ qqrobot-spring-boot-starter/src/main/java/de/honoka/qqrobot/framework/FrameworkEnum.java
@@ -1,0 +1,10 @@
+package de.honoka.qqrobot.framework;
+
+public enum FrameworkEnum {
+    
+    TESTER, 
+    
+    MIRAI, 
+    
+    ONEBOT
+}